move back markdown_browser to browser_utils
