+++ autogen/screen_parsing/test/utils/test_files/test.pptx
